assistant2: Add `Thread::use_pending_tools` method
